% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacommons-package.R
\docType{package}
\name{datacommons-package}
\alias{datacommons}
\alias{datacommons-package}
\title{datacommons: Client for the Google Data Commons API V2}
\description{
Access the Google Data Commons API V2 \url{https://docs.datacommons.org/api/rest/v2/}. Data Commons provides programmatic access to statistical and demographic data from dozens of sources organized in a knowldege graph.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidy-intelligence/r-datacommons}
  \item \url{https://tidy-intelligence.github.io/r-datacommons/}
  \item Report bugs at \url{https://github.com/tidy-intelligence/r-datacommons/issues}
}

}
\author{
\strong{Maintainer}: Christoph Scheuch \email{christoph@tidy-intelligence.com} (\href{https://orcid.org/0009-0004-0423-6819}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Teal Emery \email{lte@tealinsights.com}
}

}
\keyword{internal}
