% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacommons-package.R
\docType{package}
\name{datacommons-package}
\alias{datacommons}
\alias{datacommons-package}
\title{datacommons: Client for the Google Data Commons API V2}
\description{
An R client for accessing the Google Data Commons API V2, providing programmatic access to statistical and demographic data from various sources. This package wraps the API endpoints described at https://docs.datacommons.org/api/rest/v2/.
}
\author{
\strong{Maintainer}: Christoph Scheuch \email{your.email@example.com}

}
\keyword{internal}
