% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{dc_get_node}
\alias{dc_get_node}
\title{Retrieve Node Properties from Data Commons}
\usage{
dc_get_node(
  nodes,
  property,
  api_key = Sys.getenv("DATACOMMONS_API_KEY"),
  base_url = "https://api.datacommons.org/v2/",
  method = "list"
)
}
\arguments{
\item{nodes}{A character vector of DCIDs.}

\item{property}{A relation expression string (e.g., "<-*").}

\item{api_key}{Your Data Commons API key. If not provided, will use the
environment variable \code{DATACOMMONS_API_KEY}.}

\item{base_url}{The base URL of the Data Commons API. Defaults to the public
endpoint. For custom deployments, it must end with "/core/api/v2/".}

\item{method}{Either \code{"list"} (parsed R object) or \code{"json"} (JSON string).}
}
\value{
A list (if \code{method = "list"}) or JSON string (if \code{method = "json"}).
}
\description{
Retrieve Node Properties from Data Commons
}
\examples{
# Get all property labels for a given node
dc_get_node(nodes = "country/USA", property = "<-")

# Get one property value for a given node
dc_get_node(nodes = "dc/03lw9rhpendw5", property = "->name")

# Get multiple property values for multiple nodes
dc_get_node(
  nodes = c("geoId/06085", "geoId/06087"),
  property = "->[name, latitude, longitude]"
)

# Get all property values for a node
dc_get_node(nodes = "PowerPlant", property = "<-*")

# Get a list of all existing statistical variables
dc_get_node(nodes = "StatisticalVariable", property = "<-typeOf")

# Get a list of all existing entity types
dc_get_node(nodes = "Class", property = "<-typeOf")

}
