% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dc_set_base_url}
\alias{dc_set_base_url}
\title{Set environment variable for Data Commons base URL}
\usage{
dc_set_base_url(base_url)
}
\arguments{
\item{base_url}{description}
}
\description{
Set environment variable for Data Commons base URL
}
